+++ packages/build-app-helpers/src/index.ts
@@ -7,3 +7,4 @@ export { default as unionBy } from './unionBy';
 export { default as injectTransformRuntime } from './injectTransformRuntime';
 export { default as checkExportDefaultDeclarationExists } from './checkExportDefaultDeclarationExists';
 export { default as getSassImplementation } from './getSassImplementation';
+export { default as getLessImplementation } from './getLessImplementation';