+++ packages/plugin-app-core/src/index.ts
@@ -57,7 +57,10 @@ export default (api, options) => {
   setRegisterUserConfig(api);
 
   // register api method
-  const generator = initGenerator(api, { ...options, debugRuntime: commandArgs.debugRuntime, hasJsxRuntime });
+  if (commandArgs.debugRuntime) {
+    console.log('[deprecated] cli option --debug-runtime is deprecated, runtime file is generated as default');
+  }
+  const generator = initGenerator(api, { ...options, debugRuntime: userConfig.generateRuntime, hasJsxRuntime });
   setRegisterMethod(api, { generator });
 
   // add core template for framework