+++ packages/plugin-router/src/runtime/Router.tsx
@@ -50,7 +50,7 @@ export function parseRoutes(routes: RouteItemProps[], fallback?: React.ReactNode
   return routes.map((route) => {
     const { children, component, routeWrappers, wrappers, ...others }  = route;
     // do not wrapper components to layout added by runtime api wrapperRouteComponent
-    let mergedRouteWrappers = children ? [] : routeWrappers;
+    let mergedRouteWrappers = children ? [] : routeWrappers as IRouteWrapper[];
     if (wrappers && wrappers.length) {
       mergedRouteWrappers = mergedRouteWrappers.concat(wrappers);
     }
@@ -67,7 +67,7 @@ export function parseRoutes(routes: RouteItemProps[], fallback?: React.ReactNode
 
 export function IceRouter(props: RouterProps) {
   const { type, children, ...others } = props;
-  let renderChildren: React.ReactChild | React.ReactChildren = children;
+  let renderChildren: React.ReactChild | React.ReactChildren | undefined = children;
   if (!renderChildren && props.routes) {
     // parse routes before render
     const parsedRoutes = parseRoutes(props.routes, props.fallback);
@@ -94,34 +94,40 @@ export function Routes({ routes, fallback }: RoutesProps) {
             return <Redirect key={id} from={route.path} to={redirect} {...others} />;
           } else {
             const { component: RouteComponent, ...others } = route;
-            // React does not currently support Suspense when components are being server-side rendered
-            // process.env.__IS_SERVER__: React.RenderToString()
-            // window.__ICE_SSR_ENABLED__: React.hydrate()
-            const RenderComponent = process.env.__IS_SERVER__ || (window as any).__ICE_SSR_ENABLED__
-              ? (props: RouteComponentProps) => <RouteComponent {...props} />
-              : (props: RouteComponentProps) => {
-                return (
-                  <React.Suspense fallback={fallback || <div>loading</div>}>
-                    <RouteComponent {...props} />
-                  </React.Suspense>
-                );
-              };
+            if (RouteComponent) {
+              // React does not currently support Suspense when components are being server-side rendered
+              // process.env.__IS_SERVER__: React.RenderToString()
+              // window.__ICE_SSR_ENABLED__: React.hydrate()
+              const RenderComponent = process.env.__IS_SERVER__ || (window as any).__ICE_SSR_ENABLED__
+                ? (props: RouteComponentProps) => <RouteComponent {...props} />
+                : (props: RouteComponentProps) => {
+                  return (
+                    <React.Suspense fallback={fallback || <div>loading</div>}>
+                      <RouteComponent {...props} />
+                    </React.Suspense>
+                  );
+                };
 
-            return (
-              <Route
-                key={id}
-                {...others}
-                render={RenderComponent}
-              />
-            );
+              return (
+                <Route
+                  key={id}
+                  {...others}
+                  render={RenderComponent}
+                />
+              );
+            } else {
+              console.error('[Router] component is required when config routes');
+              return null;
+            }
           }
         } else {
           const { component: LayoutComponent, children, ...others } = route;
-          const RenderComponent = (props: RouteComponentProps) => (
+          const routesComponent = <Routes routes={children as IRouterConfig[]} fallback={fallback} />;
+          const RenderComponent = (props: RouteComponentProps) => LayoutComponent ? (
             <LayoutComponent {...props}>
-              <Routes routes={children} fallback={fallback} />
+              {routesComponent}
             </LayoutComponent>
-          );
+          ) : routesComponent;
           return (
             <Route
               key={id}