+++ packages/plugin-mpa/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+## 1.8.6
+
+- [fix] default template of mpa
+
 ## 1.8.5
 
-- fix: html template parameters of pageName
+- [fix] html template parameters of pageName